+++ Learning_Server/src/main/java/com/example/learning_server/service/Impl/AnswerServiceImpl.java
@@ -3,11 +3,14 @@ package com.example.learning_server.service.Impl;
 import com.example.learning_server.dto.request.AnswerReq;
 import com.example.learning_server.dto.request.UpdateAnswerReq;
 import com.example.learning_server.entity.Answer;
+import com.example.learning_server.exception.BusinessLogicException;
 import com.example.learning_server.mapper.AnswerMapper;
 import com.example.learning_server.repository.AnswerRepository;
 import com.example.learning_server.service.AnswerService;
+import com.example.learning_server.utils.EmailUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ObjectUtils;
 
 @Service
 @RequiredArgsConstructor
@@ -20,34 +23,45 @@ public class AnswerServiceImpl implements AnswerService {
 
     @Override
     public void addAnswer(AnswerReq answerReq) {
+        String email = EmailUtils.getCurrentUser();
+        if (ObjectUtils.isEmpty(email)) {
+            throw new BusinessLogicException();
+        }
 
         Answer answer = answerMapper.toEntity(answerReq);
-
         answerRepository.save(answer);
     }
 
     @Override
     public void updateAnswer(UpdateAnswerReq updateAnswerReq) {
-        Answer answer = answerRepository.findById(updateAnswerReq.getAnswerId()).orElse(null);
-
-        if (answer != null) {
-            if(updateAnswerReq.getContent() !=null) {
-                answer.setContent(updateAnswerReq.getContent());
-            }
-            if (updateAnswerReq.getImageLocation() != null) {
-                answer.setImageLocation(updateAnswerReq.getImageLocation());
-            }
-            if (updateAnswerReq.getVideoLocation() != null) {
-                answer.setVideoLocation(updateAnswerReq.getVideoLocation());
-            }
-            answer.setCorrect(updateAnswerReq.isCorrect());
-
-            answerRepository.save(answer);
+        String email = EmailUtils.getCurrentUser();
+        if (ObjectUtils.isEmpty(email)) {
+            throw new BusinessLogicException();
         }
+
+        Answer answer = answerRepository.findById(updateAnswerReq.getAnswerId()).orElseThrow(BusinessLogicException::new);
+
+        if (updateAnswerReq.getContent() != null) {
+            answer.setContent(updateAnswerReq.getContent());
+        }
+        if (updateAnswerReq.getImageLocation() != null) {
+            answer.setImageLocation(updateAnswerReq.getImageLocation());
+        }
+        if (updateAnswerReq.getVideoLocation() != null) {
+            answer.setVideoLocation(updateAnswerReq.getVideoLocation());
+        }
+        answer.setCorrect(updateAnswerReq.isCorrect());
+
+        answerRepository.save(answer);
     }
 
     @Override
     public void deleteAnswer(long id) {
+        String email = EmailUtils.getCurrentUser();
+        if (ObjectUtils.isEmpty(email)) {
+            throw new BusinessLogicException();
+        }
+
         answerRepository.deleteById(id);
     }
 }
